Escape backticks in Buildkite annotations
This is paranoia in case some log happens to expose a sequence of 3+
backticks that could escape the terminal code block.
