+++ src/api/buildkite/md.ts
@@ -1,0 +1,7 @@
+/**
+ * @internal
+ */
+export const md = {
+  terminal: (code: string) =>
+    `\`\`\`term\n${code.replace(/```/g, '\\`\\`\\`')}\n\`\`\``,
+};
